feat: 검색 페이지와 위치 기반 api 조회 컴포넌트 분리
